Refactor to combine dependencies into utils
