+++ lib/python/node_manager/plugins/release.py
@@ -9,7 +9,7 @@ import hou
 
 from node_manager import utils
 from node_manager.utils import definition as definition_utils
-from node_manager.dependencies import nodes
+from node_manager.utils import nodes
 
 plugin_name = "DefaultRelease"
 plugin_class = "release"
