+++ lib/python/node_manager/utils/nodes.py
@@ -10,6 +10,21 @@ import hou
 logger = logging.getLogger(__name__)
 
 
+def node_comment(current_node, published=True):
+    """Set the comment on a node.
+
+    Args:
+        current_node(hou.Node): The node to set the comment on.
+        edit(bool): Whether the node is editable or not.
+    """
+    state = "Published"
+    if not published:
+        state = "Editable"
+    current_node.setComment("Node Manager: {state}".format(state=state))
+    current_node.setGenericFlag(hou.nodeFlag.DisplayComment, True)
+    logger.debug("Set comment on node: {node}".format(node=current_node.name()))
+
+
 def node_at_path(node_path):
     """
     Get the node at a given Houdini node path.
